candidate_deliverables/fake_api/requirements.txt
flask
docker
requests
pandas
psycopg2-binary
sqlalchemy
python-dotenv
